EEPROM code added. Load Game added. Refactored.
EEPROM code added to Asokoban.ino.
Refactored LoadGame/SaveGame into loadsave.{h,cpp}.
Dropped millisecond accounting for second accounting.
Fixed bug in FmtMMSS.
Moved `GameStateStruct` to globals.h.
Added variable zoom to ardusdl.cpp.

TODO: SaveState.

+++ Asokoban/draw.h
@@ -11,7 +11,7 @@
 void DrawBoard(Piece board[HDIM][VDIM], Player worker);
 void DrawMenu(uint16_t level, uint16_t* results);
 void DrawResult(const uint8_t* text, const uint8_t row, const uint16_t
-    level, const uint16_t moves, const uint32_t elapsed);
+    level, const uint16_t moves, const uint16_t elapsed);
 void DrawStars(uint8_t set);
 //
 // vim:ft=cpp
